refactor: extract checkpoint-domain and sdlc-domain modules, move SdlcDashboardView to shared-web (Phase 6.4/6.6/6.9)
Create checkpoint-domain module with 8 pure data types and sdlc-domain module
with 13 data types, decoupling consumers from root control packages. Replace
qualified SdlcDashboardService.* type references with direct sdlc.entity imports
across views, controllers, and tests. Move SdlcDashboardView to shared-web.
